# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY API",
        "description": "API pour la gestion des comptes OMPAY",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/transactions/pay": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Effectuer un paiement",
                "description": "Permet à un client d'effectuer un paiement vers un autre client (via numéro de téléphone) ou vers un distributeur (via numéro de téléphone ou code marchand)",
                "operationId": "16f3d755bd31888ae26d626a0e08a214",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient",
                                    "type",
                                    "montant"
                                ],
                                "properties": {
                                    "recipient": {
                                        "description": "Numéro de téléphone ou code marchand du destinataire",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type de destinataire",
                                        "type": "string",
                                        "enum": [
                                            "telephone",
                                            "code_marchand"
                                        ]
                                    },
                                    "montant": {
                                        "description": "Montant du paiement",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué avec succès",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Client: John Doe\\nSolde avant: 10000 FCFA\\nDistributeur: Simonis LLC (Code: 19823394)\\nPaiement réussi!\\nSolde après: 9500 FCFA\\nTransaction créée:\\n- ID: uuid-transaction\\n- Type: paiement\\n- Montant: 500.00\\n- Code marchand: 19823394\\n- Émetteur: uuid-emetteur\\n- Bénéficiaire: uuid-beneficiaire"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Solde insuffisant"
                    },
                    "404": {
                        "description": "Destinataire introuvable",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Erreur: Code marchand '123456' introuvable."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}